requirements.txt
unstructured
tiktoken
chromadb
pypdf
openai
langchain
pandas
numpy
langchain_community
python-dotenv
wikipedia
serpapi
streamlit
logging